2015-3-9: Initial submit
A very primitive design of neuron network consisting of two reciprocally
connected layers, with 100 neurons in each layer. It is intended to
memorize and recognize some 10*10 patterns. To store a certain pattern
into the network, the same pattern should be presented 100 times.
Designed according to the paper: "Binding Sparse Spatiotemporal Patterns
In Spiking Computation"